scripts/weekly_update/requirements.txt
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
notion-client==2.4.0
pycparser==2.22
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.5
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
